separated split by sequence delimiter tests
